Reorganized testing views and  Fixed minor bug in  callGamesByDate
